Question 2, 3 et 5 ont été réglé, tout a été rajouté, sauf les case dans le switch de la classe GestionnaireEvenementServeur

+++ src/main/java/com/atoudeft/banque/OperationFacture.java
@@ -1,11 +1,11 @@
 package com.atoudeft.banque;
 
 public class OperationFacture extends Operation {
-    private double montant = 0;
+    private int montant = 0;
     private String numeroFacture;
     private String description = "";
 
-    public OperationFacture(double montant, String numeroFacture, String description) {
+    public OperationFacture(int montant, String numeroFacture, String description) {
         super.type = TypeOperation.FACTURE;
         this.montant = montant;
         this.numeroFacture = numeroFacture;
